Agent/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
clip @ git+https://github.com/openai/CLIP.git
jinja2
Pillow